removed superfluous(?) check on insertion/domain size
For some reason I was checking if the insertion box was doing odd things, and changing simMax if it was.  I think this was to do with wierd scenarios when the insertion box or simulation box were crossing 0 in that axis, but I can't see why that would be the case, so I've deleted this now.

If weird scenarios happen in future, I'll need to check this, otherwise I think we need to rely on Maya for visual checking of domain.