Update plfmaya_liggghts_part2.mel
Finally squashed a bug that's been around for a while.  For no discernible reason, when running this script for the first time upon opening Maya, the script would list objects and a couple of faces (e.g. bone1.f[764]), and this would break the entire script.  I've added -tr to line 79, so that the bones list only includes transform nodes.  Maya still selects faces, as is obvious in the viewport (selection turns orange), but the script works fine now.